Added class MultiToOneOutput that multiplexes input onto a single stream
Added class MultiToOneOutput that multiplexes several input streams onto a single stream of data and stores them in the output directory.
The size of each output file is limited by a parameter of this class which tells how many MB per output file to use.
The threads in the Compressor now takes each input file and just sends the contents through the MultiToOneOutput. 
For now no compression is done, it sends the data as is.